third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-breakpoints/set-conditional-breakpoint-expected.txt

Tests setting breakpoints.


Running: testSetConditionalBreakpointThatBreaks
Script source was shown.
Set timer for test function.
Script execution paused.
Call stack:
    0) testFunction (set-breakpoint.html:14)
Breakpoint sidebar pane 
set-breakpoint.html:14  var x = Math.sqrt(10);
Breakpoint sidebar pane 
No breakpoints
Script execution resumed.

Running: testSetConditionalBreakpointThatDoesNotBreak
Script source was shown.
Set timer for test function.
Test function finished.
Breakpoint sidebar pane 
set-breakpoint.html:14  var x = Math.sqrt(10);
Breakpoints removed.
Breakpoint sidebar pane 
No breakpoints
